Build releases with cgo disabled
For the v1.7.1 release, the runner used to build the release binaries was switching from macOS to Ubuntu. A side effect of this was that binaries would be built with CGO and dynamic linking. This can cause problems when the dynamic clib dependencies cannot be found. This commit explicitly disables CGO for static linking, as we do for all our provider plugin binaries.