Implement `coalesce` to more closely match Terraform
This PR updates our implementation of `coalesce` to more closely match
that of Terraform. Specifically, we try to convert to a most-general
type when arguments are of differing types, while respecting rules that
Terraform sets out in
https://developer.hashicorp.com/terraform/language/expressions/types#type-conversion
(e.g. booleans can lift to strings, and so on).

Part of pulumi/pulumi#18349

Co-authored-by: Brandon Pollack <brandonpollack23@gmail.com>