Bulk out tests for `distinct`
We want to start using `distinct` in `pulumi convert`. The current
implementation is likely not perfect for all cases (e.g. objects), but
it should suffice for primitive types, which hopefully will
substantially increase the completeness of `pulumi convert`. This commit
just adds a test case for booleans, and exposes the existing one for
strings that was not actually being run.

Part of pulumi/pulumi#18517