Implement the `formatlist` function (#93)
This change adds `formatlist`, which is designed to match the behaviour
of the Terraform function of the same name. `formatlist` accepts a
format string and a list of arguments, like `format`. Unlike `format`,
any lists in the set of arguments are "zipped", producing a list of
string outputs. For example, `formatlist("%s %d", ["a", "b"], [1, 2])`
yields `["a 1", "b 2"]`.

Part of pulumi/pulumi#18589