Ensure that all expected outputs appear (#91)
Currently, the test suite for this provider walks the list of actual
outputs and asserts that the values encountered match the declared
expected values. It does not however ensure that every expected value
has a corresponding actual output. As a result it's possible to add an
expectation that is not being tested, giving false confidence that
coverage is being added. This change fixes this by walking the list of
expected outputs after the fact and asserting that we saw each one of
them.