Upgrade to go-provider v1 (#102)
As part of https://github.com/pulumi/pulumi-std/pull/96 we'll need to
re-generate the SDK and I'd like to do that using a version of the CLI
that's not absolutely ancient.

Bumping our pulumi dependency doesn't work because we're on an older
version of go-provider that's no longer compatible with newer pulumi
releases (https://github.com/pulumi/pulumi-std/pull/98).

Time to rip the bandaid off and move to v1 of go-provider.

This PR is all mechanical changes:
- `infer.Function[Func, FuncArgs, FuncResponse]()` →
`infer.Function(Func{})`
- `p.Context` → `context.Context`
- `Call` → `Invoke`
- `ArgType` → `infer.FunctionRequest[ArgType]`
- `ResultType` → `infer.FunctionResult[ResultType]`
- `args.Foo` → `args.Input.Foo`
- `return ResultType{...}` → `return
infer.FunctionResponse[ResultType]{Output: ...}`